requirements.txt
deepeval==2.4.1
graphviz==0.20.3
hypernetx==2.3.13
ipython==8.12.3
langchain_core==0.3.37
langchain_openai==0.3.6
langchain_text_splitters==0.3.6
matplotlib==3.10.0
nltk==3.9.1
pandas==2.2.3
plotly==6.0.0
pydantic==2.10.6
python-dotenv==1.0.1
pyvis==0.3.2
rank_bm25==0.2.2
scikit_learn==1.6.1
scipy==1.15.2
sentence_transformers==3.4.1
torch==2.6.0
transformers==4.48.1
unstructured==0.16.17
validators==0.34.0
fastjsonschema==2.21.1
lm-format-enforcer==0.10.9
python-magic==0.4.27
unstructured[pdf]